+++ src/core/data_control_msg.rs
@@ -1,6 +1,6 @@
 use std::{fmt::Display, str::FromStr};
 
-use pareg::{key_val_arg, ArgError, FromArgStr};
+use pareg::{key_mval_arg, key_val_arg, ArgError, FromArgStr};
 use serde::{Deserialize, Serialize};
 
 use crate::{env::AppCtrl, starts};
@@ -13,10 +13,12 @@ use super::{control_msg_vec::ControlMsgVec, Error, Msg, UampApp};
 
 /// Messages that can be safely send across threads, but not necesarily esily
 /// copied.
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
 pub enum DataControlMsg {
     /// Invoke alias.
     Alias(String),
+    /// Sets the current playlist end action.
+    SetPlaylistEndAction(Option<String>),
 }
 
 impl UampApp {
@@ -27,13 +29,20 @@ impl UampApp {
         msg: DataControlMsg,
     ) -> Vec<Msg> {
         match msg {
-            DataControlMsg::Alias(name) => self
-                .config
-                .control_aliases()
-                .get(&name)
-                .map(ControlMsgVec::get_msg_vec)
-                .unwrap_or_default(),
+            DataControlMsg::Alias(name) => {
+                return self
+                    .config
+                    .control_aliases()
+                    .get(&name)
+                    .map(ControlMsgVec::get_msg_vec)
+                    .unwrap_or_default()
+            }
+            DataControlMsg::SetPlaylistEndAction(act) => {
+                self.player.playlist_mut().on_end = act;
+            }
         }
+
+        vec![]
     }
 }
 
@@ -44,6 +53,15 @@ impl FromStr for DataControlMsg {
         match s {
             v if starts!(v, "al" | "alias") => {
                 Ok(DataControlMsg::Alias(key_val_arg::<&str, _>(v, '=')?.1))
+            }
+            v if starts!(
+                v,
+                "spea" | "pl-end" | "playlist-end" | "playlist-end-action"
+            ) =>
+            {
+                Ok(DataControlMsg::SetPlaylistEndAction(
+                    key_mval_arg::<&str, _>(v, '=')?.1,
+                ))
             }
             v => Err(Error::ArgParse(ArgError::UnknownArgument(
                 v.to_owned().into(),
@@ -56,6 +74,10 @@ impl Display for DataControlMsg {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
             DataControlMsg::Alias(alias) => write!(f, "al={alias}"),
+            DataControlMsg::SetPlaylistEndAction(None) => write!(f, "spea"),
+            DataControlMsg::SetPlaylistEndAction(Some(act)) => {
+                write!(f, "spea={act}")
+            }
         }
     }
 }