+++ src/core/player/playlist.rs
@@ -6,7 +6,6 @@ use serde::{Deserialize, Serialize};
 
 use crate::{
     core::{
-        control_msg_vec::ControlMsgVec,
         library::{Library, SongId},
         SongOrder,
     },
@@ -28,7 +27,8 @@ pub struct Playlist {
     current: usize,
     #[serde(default)]
     play_pos: Option<Duration>,
-    on_end: ControlMsgVec,
+    #[serde(default)]
+    pub on_end: Option<String>,
 }
 
 impl Playlist {
@@ -45,7 +45,7 @@ impl Playlist {
             songs,
             current,
             play_pos: None,
-            on_end: ControlMsgVec::default(),
+            on_end: None,
         }
     }